refactor: 토큰 검증 로직 auth로 이동, userEntity 데코레이션 변경, lastLogin타임 저장 로직 추가
