+++ src/user/entities/user.entity.ts
@@ -8,22 +8,22 @@ export class User {
   @Column({ type: 'varchar', length: 255, unique: true, nullable: true, name: 'social_uid' })
   socialUid: string;
 
-  @Column({ type: 'varchar', length: 255, unique: true, nullable: true })
+  @Column({ type: 'varchar', length: 255, unique: true})
   email: string;
 
-  @Column({ type: 'varchar', length: 255, name: 'password_hash', nullable: true })
+  @Column({ type: 'varchar', length: 255, name: 'password_hash'})
   passwordHash: string;
 
   @Column({ type: 'varchar', length: 100 })
   name: string;
 
-  @Column({ type: 'varchar', length: 10, nullable: false })
+  @Column({ type: 'varchar', length: 10})
   gender: string;
 
-  @Column({ type: 'date', name: 'birth_date', nullable: false })
+  @Column({ type: 'date', name: 'birth_date'})
   birthDate: Date;
 
-  @Column({ type: 'varchar', length: 20, unique: true, nullable: true, name: 'phone_number' })
+  @Column({ type: 'varchar', length: 20, unique: true, name: 'phone_number' })
   phoneNumber: string;
 
   @CreateDateColumn({ type: 'timestamptz', name: 'created_at', default: () => 'CURRENT_TIMESTAMP' })
